[PRO-5532] implement campaign wizard second step - select update
