+++ ota-plus-web/app/reactapp/src/components/campaigns/wizardSteps/step2Files/index.js
@@ -1,7 +1,7 @@
-import WizardPackagesList from './PackagesList';
-import WizardPackagesListItem from './PackagesListItem';
+import SelectUpdateList from './SelectUpdateList';
+import UpdateDetails from './UpdateDetails';
 
 export {
-    WizardPackagesList,
-    WizardPackagesListItem
+    SelectUpdateList,
+    UpdateDetails,
 }
