+++ ota-plus-web/app/reactapp/src/components/campaigns/wizardSteps/step3Files/GroupsList.jsx
