+++ ota-plus-web/app/reactapp/src/components/campaigns/wizardSteps/step3Files/index.js
@@ -1,5 +1,7 @@
-import WizardPackagesVersionList from './PackagesVersionList';
+import WizardGroupsList from './GroupsList';
+import WizardGroupsListItem from './GroupsListItem';
 
 export {
-    WizardPackagesVersionList
+    WizardGroupsList,
+    WizardGroupsListItem
 }
