--- ota-plus-web/app/reactapp/src/components/campaigns/wizardSteps/step4Files/index.js
@@ -1,7 +1,0 @@
-import WizardGroupsList from './GroupsList';
-import WizardGroupsListItem from './GroupsListItem';
-
-export {
-    WizardGroupsList,
-    WizardGroupsListItem
-}
